requirements.txt
kafka-python>=2.0
psycopg2-binary>=2.8.0